ODWCPP/requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
captcha==0.3
certifi==2020.12.5
colorama==0.4.4
Django @ file:///tmp/build/80754af9/django_1613762073373/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
Pillow==8.2.0
psycopg2 @ file:///C:/ci/psycopg2_1612298715048/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
qrcode==6.1
six @ file:///C:/ci/six_1605205426665/work
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
wincertstore==0.2
